requirements.txt
git+https://github.com/aisayko/Django-tinymce-filebrowser.git
BeautifulSoup==3.2.1
Django==1.6
FeinCMS==1.7.6
PIL==1.1.7
South==0.7.6
django-filebrowser-no-grappelli==3.1.1
django-mptt==0.5.5
django-tinymce==1.5.1
feedparser==5.1.3
sorl-thumbnail==11.12
wsgiref==0.1.2
-e git+git://github.com/smacker/django-filebrowser-no-grappelli.git#egg=django-filebrowser
fabric